+++ hutool-core/src/main/java/cn/hutool/core/lang/PatternPool.java
@@ -56,6 +56,12 @@ public class PatternPool {
 	 * 移动电话
 	 */
 	public final static Pattern MOBILE = Pattern.compile("(?:0|86|\\+86)?1[3-9]\\d{9}");
+
+	/**
+	 * 18位身份证号码
+	 */
+	public final static Pattern CITIZEN_ID = Pattern.compile("[1-9]\\d{5}[1-2]\\d{3}((0\\d)|(1[0-2]))(([012]\\d)|3[0-1])\\d{3}(\\d|X|x)");
+
 	/**
 	 * 邮编
 	 */